# ioq: Implement async close() and closedir()